pyproject.toml
[tool.poetry]
name = "electronics-factory"
version = "0.1.0"
description = ""
authors = ["Azat Solonisky <asolo4150@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
ipython = "^8.27.0"
redis = "^5.0.8"
coverage = "^7.6.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.4.0"
supertools = "^1.0.1"
setuptools = "^75.1.0"
eventlet = "^0.37.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
